finish cmake reorganization
