+++ challenges/include/challenges/lists.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../../adt/include/LinkedList.hpp"
+#include "adt/LinkedList.hpp"
 
 namespace chl {
 adt::SinglyLinkedList<int> AddNumbers(adt::SinglyLinkedList<int> l1, adt::SinglyLinkedList<int> l2);
