+++ challenges/src/lists.cpp
@@ -1,4 +1,4 @@
-#include "lists.hpp"
+#include "challenges/lists.hpp"
 
 #include <utility>
