+++ challenges/test/lists.cpp
@@ -1,0 +1,3 @@
+#include "challenges/lists.hpp"
+
+#include <gtest/gtest.h>
